fs: backend: add locking
Locking is to ensure that concurrent modifications of the filesystem do
not corrupt it. For example, creating a file is a complex operation: you
have to create inode and add a directory entry that will represent it
with a path and write corresponding metadata log entries to the disk.
Simultaneous attempts of creating the same file could corrupt the file
system. Not to mention concurrent creating and unlinking of the same
path... Thus careful and robust locking mechanism is used.
For details and extensive description see src/fs/backend/shard.hh.

Signed-off-by: Krzysztof Małysa <varqox@gmail.com>