fs: compaction: add initial cluster metadata
Create a new class - data_cluster_contents_info - which holds info
about the data stored in a cluster.

Added new backend::shard members which enable finding corresponding
data_cluster_contents_info for every data cluster, as well as
discriminating the clusters that have been filled with data from ones
that are in the middle of this process.

Implemented calculations that update the data stored in those metadata
log fields.

Signed-off-by: Wojciech Mitros <wmitros@protonmail.com>