fs: compaction: data compaction execution
The decision on which clusters to compact and when should it be done is
based on two variables: "compactness" and "max_data_compaction_memory",
which are set as parameters of the filesystem. The compaction is
performed when enough clusters contain little up-to-date data; enough
clusters meaning so much that including another one would require using
more memory than max_data_compaction_memory, and little data meaning
less than compactness*size of cluster.
This patch adds those variables and the decision making that is based
on them.

Signed-off-by: Wojciech Mitros <wmitros@protonmail.com>