fs: backend: add read operation
Reads file data from disk and memory based on information stored in
inode's data vectors. When on-disk data and given buffer are
properly aligned then read disk data is not stored in a temporary
buffer but is directly read into the buffer given by the caller.

Added device_reader to perform unaligned reads with caching.

Signed-off-by: Michał Niciejewski <quport@gmail.com>