fs: backend: add metadata_log::to_disk_buffer
To disk buffers buffer data that will be written to disk. There are two
kinds: (normal) to disk buffer and metadata to disk buffer. The latter
is implemented using the former, but provides higher level interface for
appending metadata log entries rather than raw bytes.

Normal to disk buffer appends data sequentially, but if a flush occurs
the offset where next data will be appended is aligned up to alignment
to ensure that writes to the same cluster are non-overlapping.

Metadata to disk buffer appends data using normal to disk buffer but
does some formatting along the way. The structure of the metadata log on
disk is as follows:
| checkpoint_1 | entry_1, entry_2, ..., entry_n | checkpoint_2 | ... |
               | <---- checkpointed data -----> |
etc. Every batch of metadata_log entries is preceded by a checkpoint
entry. Appending metadata log appends the current batch of entries.
Flushing or lack of space in the current cluster ends current batch of
entries and then checkpoint entry is updated (because it holds CRC code
of all checkpointed data) and then write of the whole batch is requested
and a new checkpoint (if there is space for that) is started.
Last checkpoint in a cluster contains a special entry pointing to the
next cluster that is utilized by the metadata log.

Signed-off-by: Krzysztof Małysa <varqox@gmail.com>