http: throw on duplicated routes to prevent memory leaks
The current interface of adding exact HTTP routes implicitly
allows memory leaks, since it expects the callee to allocate
the handler, but then offers to override the handler without
freeing it. In order to prevent users from leaking memory,
an exception is thrown when a duplicated handler is registered,
and the user-side solution is to simply remove the old handler first.
Alternatively, routes::put could call `delete` by itself,
but that sounds error-prone - overriding an existing route
is most likely programmer's oversight, not a conscious decision -
and if it is conscious, it's enough to just drop the old handler first.

This introduces a minor API change, but since routes::put()
was never marked as noexcept, it does not require bumping
the API version.

Note1: the API is a little bit inconsistent, since routes::drop()
returns a pointer for a user to delete manually, but at the same time
~routes destructor frees the pointers on its own. Nevertheless,
that's the current API and this commit does not change it.

Note2: it would be nice to offer an additional API layer without
raw pointers, but that should be a separate series, since it
would involve introducing new API functions and possibly deprecating
the old ones.