abort_source: make the subscription noexcept
As the FIXME suggested, the subscription function is now required
to be noexcept.
Verified by manually injecting a throw in a test case, which resulted
in terminating the app with std::terminate.
All (one) internal Seastar callers are verified as noexcept:
 - subscribing a timer cancellation.

The FIXME was likely created because the callback is later used to
dispose of elements in boost's intrusive list, which expects
it to not throw (or, the behaviour becomes undefined, because
the dispose loop is interrupted). Here's boost code with the comment:
   //! <b>Requires</b>: Disposer::operator()(pointer) shouldn't throw.
   //!
   //! <b>Effects</b>: Erases all the elements of the container.
   //!   No destructors are called.
   //!   Disposer::operator()(pointer) is called for the removed elements.
   //!
   //! <b>Throws</b>: Nothing.
   //!
   //! <b>Complexity</b>: Linear to the number of elements of the list.
   //!
   //! <b>Note</b>: Invalidates the iterators to the erased elements.
   template <class Disposer>
   void clear_and_dispose(Disposer disposer)
   {
      const_iterator it(this->begin()), itend(this->end());
      while(it != itend){
         node_ptr to_erase(it.pointed_node());
         ++it;
         if(safemode_or_autounlink)
            node_algorithms::init(to_erase);
         disposer(get_real_value_traits().to_value_ptr(to_erase));
      }
      node_algorithms::init_header(this->get_root_node());
      this->priv_size_traits().set_size(0);
   }

NOTE: The alternative solution is to wrap the implementation
of abort_source::subscription::on_abort() with a try-catch clause,
but that's potentially dangerous, since maybe the exception
is not recoverable and should not be ignored.
Thus, it becomes caller's responsibility to provide a noexcept
callback, under the assumption that std::terminate is better than UB.