apps/fs_perf: added sfs_perf application
sfs_perf is an application created to measure the performance of
SeastarFS in scenarios resembling the use of filesystem by a database.

The application creates some specified number of files. Next, it starts
randomly performing operations on these files: reading, writing or
truncating until it exceeds some specified at application startup
amount of work (for example some number of operations). Also, to cope
with out-of-space problems, when approximate used file system disk
space exceeds 50% then series of truncates is started to decrease this
level to 25%.

One can set the test specification by passing command-line options to
the application. All options, with short descriptions, can be found by
starting the application with --help option.

Signed-off-by: Michał Niciejewski <quport@gmail.com>